cleanup mixin part 2
