+++ addon/mixins/frost-list-core-mixin.js
@@ -1,0 +1,33 @@
+import Ember from 'ember'
+const {Mixin, on} = Ember
+import FrostListSelectionMixin from 'ember-frost-list/mixins/frost-list-selection-mixin'
+import FrostListExpansionMixin from 'ember-frost-list/mixins/frost-list-expansion-mixin'
+import FrostListSortingMixin from 'ember-frost-list/mixins/frost-list-sorting-mixin'
+import computed from 'ember-computed-decorators'
+import createActionClosure from 'ember-frost-list/utils/action-closure'
+
+export default Mixin.create({
+  initListCoreMixin: on('init', function () {
+    Ember.defineProperty(this, '_listItems', Ember.computed.alias(this.listConfig.items))
+  }),
+
+  @computed('_listItems.[]')
+  listItems (listItems) {
+    let wrapper = []
+    return listItems.map((item) => {
+      return wrapper.pushObject(Ember.Object.create({
+        id: item.id,
+        record: item
+      }))
+    })
+  },
+
+  @computed('listItems.[]', 'selectedItems', 'expandedItems')
+  StatefulListItems (listItems, selectedItems, expandedItems) {
+    return listItems.map((item) => {
+      item.set('isSelected', selectedItems.getWithDefault(item.id, false))
+      item.set('isExpanded', expandedItems.getWithDefault(item.id, false))
+      return item
+    })
+  }
+})
