+++ addon/mixins/frost-list-expansion-mixin.js
@@ -3,8 +3,10 @@ const {
   Mixin,
   on
 } = Ember
+import FrostListCoreMixin from 'ember-frost-list/mixins/frost-list-core-mixin'
 
-export default Mixin.create({
+
+export default Mixin.create(FrostListCoreMixin, {
   // == Event =================================================================
   initListExpansionMixin: on('init', function () {
     this.set('expandedItems', Ember.Object.create())
