+++ addon/mixins/frost-list-mixin.js
@@ -8,8 +8,6 @@ import createActionClosure from 'ember-frost-list/utils/action-closure'
 
 export default Mixin.create(FrostListSelectionMixin, FrostListExpansionMixin, FrostListSortingMixin, {
   initListMixin: on('init', function () {
-    Ember.defineProperty(this, '_listItems', Ember.computed.alias(this.listConfig.items))
-
     // create closures
     Ember.defineProperty(this, '_selectItem', undefined,
       createActionClosure.call(this, this.actions.selectItem)
@@ -35,26 +33,6 @@ export default Mixin.create(FrostListSelectionMixin, FrostListExpansionMixin, Fr
       createActionClosure.call(this, this.actions.sortItems)
     )
   }),
-
-  @computed('_listItems.[]')
-  listItems (listItems) {
-    let wrapper = []
-    return listItems.map((item) => {
-      return wrapper.pushObject(Ember.Object.create({
-        id: item.id,
-        record: item
-      }))
-    })
-  },
-
-  @computed('listItems.[]', 'selectedItems', 'expandedItems')
-  StatefulListItems (listItems, selectedItems, expandedItems) {
-    return listItems.map((item) => {
-      item.set('isSelected', selectedItems.getWithDefault(item.id, false))
-      item.set('isExpanded', expandedItems.getWithDefault(item.id, false))
-      return item
-    })
-  },
 
   listMixinConfig: Ember.computed('activeSorting', 'sortableProperties', function () {
     let activeSorting = this.get('activeSorting')
