+++ addon/mixins/frost-list-selection-mixin.js
@@ -1,45 +1,20 @@
 import Ember from 'ember'
 const {Mixin, on} = Ember
+import {updateSelectedItemsHash} from 'ember-frost-list/utils/utils'
+import FrostListCoreMixin from 'ember-frost-list/mixins/frost-list-core-mixin'
 
-export default Mixin.create({
+
+export default Mixin.create(FrostListCoreMixin, {
   // == Event =================================================================
   initListSelectionMixin: on('init', function () {
     this.set('selectedItems', Ember.Object.create())
   }),
 
-  // == Functions ==============================================================
-  updateSelectedItemsHash (selections, attrs) {
-    let _selections = selections
-    if (attrs.selectDesc.isSelected) {
-      if (attrs.selectDesc.isShiftSelect) {
-        attrs.records.forEach((record) => {
-          _selections.set(record.id, true)
-        })
-      } else {
-        if ((!attrs.selectDesc.isTargetSelectionIndicator && !attrs.selectDesc.isCtrlSelect)) {
-          Object.keys(_selections).forEach((key) => {
-            _selections.set(key, false)
-          })
-        }
-        attrs.records.forEach((record) => {
-          _selections.set(record.id, true)
-        })
-      }
-    } else {
-      attrs.records.forEach((record) => {
-        _selections.set(record.id, false)
-      })
-    }
-    return _selections
-  },
-
-  // == Computed Properties ====================================================
-
   // == Actions ================================================================
   actions: {
     selectItem (attrs) {
       let selectedItems = this.get('selectedItems')
-      this.set('selectedItems', this.updateSelectedItemsHash(selectedItems, attrs))
+      this.set('selectedItems', updateSelectedItemsHash(selectedItems, attrs))
       this.notifyPropertyChange('selectedItems')
     }
   }
