+++ addon/mixins/frost-list-sorting-mixin.js
@@ -4,8 +4,10 @@ const {
   on
 } = Ember
 import computed from 'ember-computed-decorators'
+import FrostListCoreMixin from 'ember-frost-list/mixins/frost-list-core-mixin'
 
-export default Mixin.create({
+
+export default Mixin.create(FrostListCoreMixin, {
   // == Event =================================================================
   initListSortingMixin: on('init', function () {
     this.set('queryParams', ['listConfig.sorting.active'])
